requirements.txt
msqlclient
django
djangorestframework


qrcode
pillow 

